% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHDL.optim.R
\name{sHDL.optim}
\alias{sHDL.optim}
\title{Maximize the log likelihood to infer the heritability enrichment fold(s).}
\usage{
sHDL.optim(
  ref.data,
  N,
  start.v = NULL,
  output.file = NULL,
  log.file = "",
  stepwise = TRUE,
  fix.h2 = NULL,
  fix.intercept = NULL,
  fix.eta = NULL,
  lim = exp(-18),
  verbose = FALSE,
  lwr = NULL,
  upr = NULL,
  maxit = 1000,
  pgtol = 0.001,
  par.h2 = FALSE,
  mc.cores = 1,
  nthreads = 1
)
}
\arguments{
\item{ref.data}{A list of data matched to LD reference generated by \code{\link{sHDL.joint.reduct.dim}}.}

\item{N}{The sample size of the GWAS.}

\item{start.v}{Starting values for \code{c(h2, intercept, eta, fold1, fold2, ..., foldn)}, where n is the number of annotations. Default is \code{NULL}, which means \code{c(0.1, 1, 0, 1, ..., 1)}. If \code{fix.h2}, \code{fix.intercept} or \code{fix.eta} is not \code{NULL}, the starting values for \code{h2}, \code{intercept} or \code{eta} will be ignored.}

\item{output.file}{Where the results should be written.}

\item{log.file}{Where the log should be written.}

\item{stepwise}{Whether to estimate enrichment fold by estimating heritability and intercept first, default \code{stepwise = FALSE}. If \code{fix.h2}, \code{fix.intercept} and \code{fix.eta} are not \code{NULL}, \code{stepwise} will be overridden.}

\item{fix.h2}{Whether to fix the heritability to \code{fix.h2} or estimate the heritability, default \code{fix.h2 = NULL}, which means estimate the heritability.}

\item{fix.intercept}{Whether to fix the intercept to \code{fix.intercept} or estimate the intercept, default \code{fix.intercept = NULL}, which means estimate the intercept.}

\item{fix.eta}{Whether to fix the eta to \code{fix.eta} or estimate the eta, default \code{fix.eta = NULL}, which means estimate the eta. If \code{fix.eta} is not \code{NULL}, the eta will be fixed to \code{fix.eta}.}

\item{lim}{Tolerance limitation to ensure the positive-definiteness of covariance matrices, default \code{lim = exp(-18)}.}

\item{verbose}{Whether to print the log on the console, default \code{verbose = FALSE}.}

\item{lwr}{Lower bounds for \code{c(h2, intercept, eta, fold1, fold2, ..., foldn)}. Default is \code{NULL}, which means \code{c(0, 0, -2, 0, 0, ..., 0)}. When \code{fix.h2}, \code{fix.intercept} or \code{fix.eta} is not \code{NULL}, the lower bounds for \code{h2}, \code{intercept} or \code{eta} will be ignored.}

\item{upr}{Upper bounds for \code{c(h2, intercept, eta, fold1, fold2, ..., foldn)}, default is \code{NULL}, which means \code{c(1, 5, 1, M/Md_1, M/Md_2, ..., M/Md_n)}, where \code{Md_i} is the sum of the weights of the i-th annotation. When \code{fix.h2}, \code{fix.intercept} or \code{fix.eta} is not \code{NULL}, the upper bounds for \code{h2}, \code{intercept} or \code{eta} will be ignored.}

\item{maxit}{Maximum number of iterations, default maxit = 1000.}

\item{pgtol}{Tolerance for convergence, default pgtol = 1e-3.}

\item{par.h2}{Whether to estimate the partitioned heritability, default \code{par.h2 = FALSE}.}

\item{mc.cores}{Number of cores to use for parallelization, default \code{mc.cores = 1}.}

\item{nthreads}{Number of threads to use for matrix operations, default \code{nthreads = 1}. The default value is suitable for most cases, do not change it unless you are sure about the performance.}

\item{annots}{A vector of annotation names, default \code{NULL}, which means all annotations in the \code{Dr.path} will be used.}

\item{lam.cut}{Cutoff for the eigenvalues, default is \code{NULL}, which means no cutoff. For analyses with a limited number of traits and annotations, a lower cutoff (such as 0.1, or even not using a cutoff at all) is recommended. For large-scale analyses, a higher cutoff (such as 1)  is recommended, to yield fast computation.}
}
\value{
A data.frame is returned with:
\itemize{
\item{item } The name of the parameter.
\item{estimation } The estimated value of the parameter.
\item{se } The standard error of the parameter.
\item{p } The p-value of the parameter.
\item{note } The note of the parameter.
}
}
\description{
Maximize the log likelihood to infer the heritability enrichment fold(s).
}
