% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHDL.rebuild.ref.R
\name{sHDL.rebuild.ref}
\alias{sHDL.rebuild.ref}
\title{Re-build the LD reference panel with given SNPs.}
\usage{
sHDL.rebuild.ref(
  LD.path,
  gwas.snps,
  LD.path.new,
  nthreads = 1,
  lam.cut = NULL,
  pattern = ".*chr(\\\\d{1,2})\\\\.(\\\\d{1,2})[_\\\\.].*"
)
}
\arguments{
\item{LD.path}{Path to the \code{.rda} file where the Eigen decomposition of LD matrix is stored.}

\item{gwas.snps}{A vector contains SNP IDs from the GWAS summary statistics.}

\item{LD.path.new}{The new path to the re-built LD reference.}

\item{nthreads}{Number of threads to use, default \code{nthreads = 1}.}

\item{lam.cut}{Eigenvalue cutoff for LD matrices, default \code{lam.cut = NULL}, which means the same cutoff in the original LD reference panel. For analyses with a limited number of traits and annotations, a lower cutoff (such as 0.1, or even not using a cutoff at all) is recommended. For large-scale analyses, a higher cutoff (such as 1)  is recommended, to yield fast computation.}

\item{pattern}{Chromosome and picece pattern of LD files, default is \code{".*chr(\\d{1,2})\\.(\\d{1,2})[_\\.].*"}.}
}
\value{
The new path to the re-built LD reference.
}
\description{
Re-build the LD reference panel with given SNPs.
}
