% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHDL.optim.R
\name{sHDL.optim}
\alias{sHDL.optim}
\title{Maximize the log likelihood to infer the heritability enrichment fold.}
\usage{
sHDL.optim(
  ref.data,
  N,
  start.v = c(1, 0.1, 1),
  output.file = NULL,
  log.file = "",
  stepwise = FALSE,
  fix.h2 = NULL,
  fix.intercept = NULL,
  lim = exp(-18),
  verbose = FALSE,
  clust = NULL,
  lwr = NULL,
  upr = NULL,
  maxit = 1000,
  pgtol = 0.001
)
}
\arguments{
\item{ref.data}{A list of data matched to LD reference generated by \code{\link{sHDL.reduct.dim}}.}

\item{N}{The sample size of the GWAS.}

\item{start.v}{A vector of starting values \code{c(fold, h2, intercept)} for optimization.}

\item{output.file}{Where the log and results should be written.}

\item{log.file}{Where the log should be written.}

\item{stepwise}{Whether to estimate enrichment fold by estimating heritability and intercept first, default \code{stepwise = FALSE}. If \code{fix.h2} and \code{fix.intercept} are not \code{NULL}, \code{stepwise} will be overridden.}

\item{fix.h2}{Whether to fix the heritability to \code{fix.h2} or estimate the heritability, default \code{fix.h2 = NULL}, which means estimate the heritability.}

\item{fix.intercept}{Whether to fix the intercept to \code{fix.intercept} or estimate the intercept, default \code{fix.intercept = NULL}, which means estimate the intercept.}

\item{lim}{Tolerance limitation to ensure the positive-definiteness of covariance matrices, default \code{lim = exp(-18)}.}

\item{verbose}{Whether to print the log on the console, default \code{verbose = FALSE}.}

\item{clust}{A cluster object generated by \code{\link{parallel::makeCluster}}.}

\item{lwr}{Lower bounds for \code{c(fold, h2, intercept)}. Default is \code{NULL}, which means \code{c(0, 0, 0.1)}.}

\item{upr}{Upper bounds for \code{c(fold, h2, intercept)}. Default is \code{NULL}, which means \code{c(M / Md, 1, 5)}, where \code{Md} is the sum of annotation weights and \code{M} is the total number of SNPs.}

\item{maxit}{Maximum number of iterations, default maxit = 1000.}

\item{pgtol}{Tolerance for convergence, default pgtol = 1e-3.}
}
\value{
A list is returned with:
\itemize{
\item{time }{Time elapsed in seconds for optimization.}
\item{fold }{The estimated heritability enrichment fold.}
\item{h2 }{The estimated SNP-based heritability.}
\item{intercept }{The estimated intercept.}
\item{fold.se }{The standard error of the estimated heritability enrichment fold.}
\item{h2.se }{The standard error of the estimated heritability.}
\item{intercept.se }{The standard error of the estimated intercept.}
\item{fold.p }{P-value based on Wald test for the estimated heritability enrichment fold.}
\item{h2.p }{P-value based on Wald test for the estimated heritability.}
\item{intercept.p }{P-value based on Wald test for the estimated intercept.}
\item{stepwise }{Whether the optimization is done in a stepwise manner.}
\item{converged }{Whether the optimization converges.}
\item{message }{The message returned by \code{\link{optim}}.}
}
}
\description{
Maximize the log likelihood to infer the heritability enrichment fold.
}
