% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHDL.R
\name{sHDL}
\alias{sHDL}
\title{Stratified High-Definition Likelihood (sHDL) Inference of Heritability Enrichment}
\usage{
sHDL(
  D,
  gwas.df,
  LD.path,
  output.file = NULL,
  nthreads = 1,
  stepwise = TRUE,
  fill.missing.N = NULL,
  lim = exp(-18),
  lam.cut = NULL,
  Dr.path = "./Dr",
  overwrite = FALSE,
  verbose = FALSE,
  fix.h2 = NULL,
  fix.intercept = NULL,
  maxit = 1000,
  pgtol = 0.001,
  start.v = c(1, 0.1, 1),
  mode = c("disk", "memory"),
  pattern = ".*chr(\\\\d{1,2})\\\\.(\\\\d{1,2})[_\\\\.].*",
  norm.method = c("minmax", "scaled", "none")
)
}
\arguments{
\item{D}{A vector of genomic annotations with vector names of SNP IDs.}

\item{gwas.df}{A data frame including GWAS summary statistics of genetic variants for a trait.
The input data frame should include following columns: SNP, SNP ID; A1, effect allele; A2, reference allele;
N, sample size; Z, z-score; If Z is not given, alternatively, you may provide: b, estimate of marginal effect in GWAS; se, standard error of the estimates of marginal effects in GWAS.}

\item{LD.path}{Path to the directory where linkage disequilibrium (LD) information is stored, compatible with HDL LD reference panels.}

\item{output.file}{Where the log and results should be written. If you do not specify a file, the log will be printed on the console.}

\item{nthreads}{Number of threads to use, default \code{nthreads = 1}.}

\item{stepwise}{Whether to estimate enrichment fold by estimating heritability and intercept first, default \code{stepwise = FALSE}. If \code{fix.h2} and \code{fix.intercept} are not NULL, stepwise will be overridden.}

\item{fill.missing.N}{If the sample size is missing in the GWAS summary statistics, fill.missing.N is used to fill the missing values.}

\item{lim}{Tolerance limitation to ensure the positive-definiteness of covariance matrices, default \code{lim = exp(-18)}.}

\item{lam.cut}{Eigenvalue cutoff for LD matrices, default \code{lam.cut = NULL}, which means no cutoff.}

\item{Dr.path}{Path to the directory where the Dr matrices are stored, default \code{Dr.path = "./Dr"}.}

\item{overwrite}{Whether to overwrite the existing Dr matrices, default \code{overwrite = FALSE}.}

\item{verbose}{Whether to print the log on the console, default \code{verbose = FALSE}.}

\item{fix.h2}{Whether to fix the heritability to \code{fix.h2} or estimate the heritability, default \code{fix.h2 = NULL}, which means estimate the heritability.}

\item{fix.intercept}{Whether to fix the intercept to \code{fix.intercept} or estimate the intercept, default \code{fix.intercept = NULL}, which means estimate the intercept.}

\item{maxit}{Maximum number of iterations, default \code{maxit = 1000}.}

\item{pgtol}{Tolerance for convergence, default \code{pgtol = 1e-3}.}

\item{start.v}{A vector of starting values \code{c(fold, h2, intercept)} for optimization.}

\item{mode}{Whether to store Dr to disk or memory, default \code{mode = "disk"}. If \code{mode = "disk"}, \code{Dr} is stored to disk (path returned only) and lam are not returned. If \code{mode = "memory"}, \code{Dr} and \code{lam} are returned.}

\item{pattern}{Chromosome and picece pattern of LD files, default is \code{".*chr(\\d{1,2})\\.(\\d{1,2})[_\\.].*"}.}

\item{norm.method}{The normalization method, either \code{"minmax"} (default), \code{"scaled"} or \code{"none"}. If \code{"minmax"}, the annotation weight vector D is normalized to [0, 1]. If \code{"scaled"}, the sum of normalized vector D is scaled to the number of annotated SNPs. If \code{"none"}, the annotation weight vector D is not normalized.}
}
\value{
A list is returned with:
\itemize{
\item{time }{The time used for optimization.}
\item{fold }{The estimated heritability enrichment fold.}
\item{h2 }{The estimated SNP-based heritability.}
\item{intercept }{The estimated intercept.}
\item{fold.se }{The standard error of the estimated heritability enrichment fold.}
\item{h2.se }{The standard error of the estimated heritability.}
\item{intercept.se }{The standard error of the estimated intercept.}
\item{fold.p }{P-value based on Wald test for the estimated heritability enrichment fold.}
\item{h2.p }{P-value based on Wald test for the estimated heritability.}
\item{intercept.p }{P-value based on Wald test for the estimated intercept.}
\item{stepwise }{Whether the optimization is done in a stepwise manner.}
\item{converged }{Whether the optimization converges.}
\item{message }{The message returned by \code{\link{optim}}.}
}
}
\description{
The function return estimation and standard error of heritability enrichment fold of trait based on GWAS summary statistics.
}
\note{
Users can download the precomputed eigenvalues and eigenvectors of LD correlation matrices for European ancestry population. The download link can be found at https://github.com/zhenin/HDL/wiki/Reference-panels
These are the LD matrices and their eigen-decomposition from 335,265 genomic British UK Biobank individuals. Three sets of reference panel are provided:
1) 1,029,876 QCed UK Biobank imputed HapMap3 SNPs. The size is about 33 GB after unzipping. Although it takes more time, using the imputed panel provides more accurate estimates of genetic correlations.
Therefore if the GWAS includes most of the HapMap3 SNPs, then we recommend using the imputed reference panel.
2) 769,306 QCed UK Biobank imputed HapMap2 SNPs. The size is about 18 GB after unzipping.If one of your GWAS includes most of the HapMap 2 SNPs, but many SNPs (more than 1%) in the above HapMap 3 reference panel are absent,
then this HapMap2 panel is more proper to be used for HDL.
3) 307,519 QCed UK Biobank Axiom Array SNPs. The size is about 7.5 GB after unzipping.
}
\examples{
\dontrun{
## The GWAS summary statistics for birth weight loaded from HDL package.
data(gwas1.example, package='HDL')
M <- nrow(gwas1.example)
set.seed(1234)
D <- rbinom(M, 1, 0.01) # random D vector
names(D) <- gwas1.example$SNP

## The path to the directory where linkage disequilibrium (LD) information is stored.
LD.path <- "path/to/UKB_array_SVD_eigen90_extraction"

## To speed up the test, we set a large lam.cut value.
res.sHDL <- sHDL(D, gwas1.example, LD.path, nthreads=1, stepwise=TRUE, lam.cut=10, Dr.path="./Dr", verbose=F)
as.data.frame(res.sHDL)
system("rm -rf ./Dr") # remove the Dr directory
}
}
\references{
Lan A and Shen X (2024). Modeling the Genetic Architecture of Complex Traits via Stratified High-Definition Likelihood.
}
\author{
Ao Lan, Xia Shen
}
