% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHDL.reduct.dim.R
\name{sHDL.reduct.dim}
\alias{sHDL.reduct.dim}
\title{Stratified High-Definition Likelihood (sHDL) Inference of Heritability Enrichment}
\usage{
sHDL.reduct.dim(
  LD.path,
  z = NULL,
  D = NULL,
  lam.cut = NULL,
  Dr.path = NULL,
  overwrite = FALSE,
  mode = c("disk", "memory"),
  nthreads = 1,
  pattern = ".*chr(\\\\d{1,2})\\\\.(\\\\d{1,2})[_\\\\.].*"
)
}
\arguments{
\item{z}{A matrix of Z-scores with rownames of SNP IDs. Supporting multiple columns for multiple traits.}

\item{D}{A vector of genomic annotations with vector names of SNP IDs.}

\item{lam.cut}{Eigenvalue cutoff for LD matrices, default lam.cut = NULL, which means no cutoff.}

\item{Dr.path}{Path to the directory where the Dr matrices are stored, default Dr.path = NULL, which means do not store Dr to disk.}

\item{overwrite}{Whether to overwrite the existing Dr matrices, default overwrite = FALSE.}

\item{mode}{Whether to store Dr to disk or memory, default \code{mode = "disk"}. If \code{mode = "disk"}, \code{Dr} is stored to disk (path returned only) and lam are not returned. If \code{mode = "memory"}, \code{Dr} and \code{lam} are returned.}

\item{nthreads}{Number of threads to use for matrix operations, default \code{nthreads = 1}.}

\item{pattern}{Chromosome and picece pattern of LD files, default is \code{".*chr(\\d{1,2})\\.(\\d{1,2})[_\\.].*"}.}

\item{LD.file}{Path to the \code{.rda} file where the Eigen decomposition of LD matrix is stored.}
}
\value{
A list is returned with:
\itemize{
\item{Dr }{The reduct RDR matrix.}
\item{zr }{The reduct z-score vector.}
\item{lam }{The eigenvalues of LD matrix.}
\item{Md }{The sum of annotation weights of SNPs in given LD matrix.}
\item{M }{The number of SNPs in given LD matrix.}
}
}
\description{
The function reduces the dimension of covariance matrix by converting z (D) to zr (Dr).
}
